pyproject.toml
[project]
name = "claender-mcp"
version = "0.1.0"
description = "Google Calendar MCP Server — FastAPI + Claude NLP + Docker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "mcp>=1.9.0",
    "pytz>=2024.1",
    "structlog>=24.0.0",
]
